New CLI and old ckan command needs to be side by side
