+++ ckanext/xroad_integration/utils.py
@@ -1,0 +1,4 @@
+def init_db():
+    import ckan.model as model
+    from ckanext.xroad_integration.model import init_table
+    init_table(model.meta.engine)
